move C++ side of birgen to bir_codegen
